Create index.html
This is part of a larger project where my website displays social media both at the top and bottom of my webpage. The social media icons are an easy way to access my social accounts for: linkedin, Github, YouTube, Twitter, Facebook, Pinterest and Instagram. The AddThis.com site was used to generate the social plugins and contains analytic statistics for my website as part of my registration. The code was copied from AddThis.com with styling and javascript file attached.