Create site.js
The AddThis.com site did not include GitHub into it's family of social icons, i.e. the minified javascript file copied over. I was fortunate enough to be able to assume that styling would be provided as part of the plugin and it was, however functionality did not exist for an onclick event so this had to be added to direct the user to the href added here.