Create site.less
The background-position is the placement of the new greyscale icons according to 7 icons on top of each other 32x32 pixels with a width of 32 pixels and a height of 224 pixels and here placed with changing ypos of height of above icon. The !important is to ensure that the original icon images are overridden according to the style sheet provided for the plugin from the AddThis.com site.

The classes required to style are evident upon element inspect in the developer tools and not before as the style sheet applied is minified.